Refactoring details
