+++ mobile-app/app/src/main/java/com/mobile/test/model/Login/LoginResponse.kt
@@ -1,4 +1,4 @@
-package com.mobile.test.model
+package com.mobile.test.model.Login
 
 data class LoginResponse(
     val token: String
